// src/app/models/users.interface.ts
export interface Users {
    id: string;
    position: string;
    office: string;
    salary: number;
    email: string;
    firstName: string;
    lastName: string;

    // constructor(position: string, office: string, salary: number, email: string, firstName: string, lastName: string) {
    //     // this.id = id;
    //     this.position = position;
    //     this.office = office;
    //     this.salary = salary;
    //     this.email = email;
    //     this.firstName = firstName;
    //     this.lastName = lastName;
    // }
}